+++ src/scielo/bin/xml/config/collection.xmlproc.ini
@@ -1,39 +1,40 @@
-CISIS_PATH=/bases/xml.000/proc/cisis
-JAVA_PATH=/usr/local/jre1.5.0_06/bin
+PATH_CISIS_1030=
+PATH_CISIS_1660=
 
-DB_ISSUE_FILENAME=/bases/xml.000/serial_proc/issue/issue
+WEB_PATH=
+SERIAL_PATH=
+ISSUE_DB=
+ISSUE_DB_COPY=
 
+FTP_SERVER=
+FTP_USER=
+FTP_PASSWORD=
+FTP_DIR=
 
+EMAIL_SERVICE=on
+SENDER_NAME=
+SENDER_EMAIL=
+EMAIL_TO=
+DOWNLOAD_PATH=
+TEMP_PATH=
+QUEUE_PATH=
 
-COL_PROC_SERIAL_PATH=/bases/xml.000/xmldata/col/scl/4proc/serial
-
-XML_PATH=/bases/xml.000/xmldata/col/scl/4web/bases/xml
-PDF_PATH=/bases/xml.000/xmldata/col/scl/4web/bases/pdf
-IMG_PATH=/bases/xml.000/xmldata/col/scl/4web/htdocs/img
-COL_SCILISTA=/bases/xml.000/xmldata/col/scl/4proc/serial/scilista.lst
-
-DOWNLOAD_PATH=/bases/xml.000/xmldata/col/scl/downloads
-DOWNLOAD_ARCHIVE_PATH=/bases/xml.000/xmldata/col/scl/4archive
-WORK_PATH=/bases/xml.000/xmldata/col/scl/work
-
-SENDER_NAME=SciELO Brasil
-SENDER_EMAIL=roberta.takenaka@scielo.org
-BCC_EMAIL=liana.holanda@scielo.org,viviane.pereira@scielo.org,renato.galdino@scielo.org,eliana.salgado@scielo.org,suely.lucena@scielo.org,caroline.soares@scielo.org,jesner.santos@scielo.org,renata.almeida@scielo.org,javani.araujo@scielo.org,leticia.aquino@scielo.org,roberta.takenaka@scielo.org
-
-IS_AVAILABLE_EMAIL_SERVICE=yes
-
-EMAIL_SUBJECT_PREFIX=[SciELO-XML] [Brasil] XML evaluation report of  
+EMAIL_SUBJECT=[SciELO-XML] [Brasil] XML evaluation report of  
 EMAIL_TEXT=email.txt
 
-EMAIL_SUBJECT_PREFIX_DOWNLOAD=[SciELO-XML] [Brasil] Download of XML files packages 
-EMAIL_TEXT_DOWNLOAD=/bases/xml.000/xmlproc/email_download.txt
+EMAIL_SUBJECT_DOWNLOAD=[SciELO-XML] [Brasil] Download of XML files packages 
+EMAIL_TEXT_DOWNLOAD=email_download.txt
 
-EMAIL_SUBJECT_WORK_PATH=[SciELO-XML] [Brasil] Packages not processed
-EMAIL_TEXT_WORK_PATH=/bases/xml.000/xmlproc/email_invalid_packages.txt
+EMAIL_SUBJECT_NOT_PROCESSED=[SciELO-XML] [Brasil] Packages not processed
+EMAIL_TEXT_NOT_PROCESSED=email_invalid_packages.txt
 
-FTP_SERVER=ftp.scielo.br
-FTP_USER=scielo
-FTP_PSWD=sci2003
-FTP_DIR=_XML/scielo.br
+DOWNLOAD_PATH=
+TEMP_PATH=
+QUEUE_PATH=
+DOWNLOAD_ARCHIVE_PATH=
 
+XML_PATH=
+PDF_PATH=
+IMG_PATH=
+COL_SCILISTA=
