+++ src/scielo/bin/xml/modules/issue.fst
@@ -1,0 +1,7 @@
+1 0 mpl,v35,if (v32='ahead' or v32='review') and val(v65*0.4)>2007 then v65*0.4 fi,|v|v31,|s|v131,|n|v32,|s|v132,v41/,v35,v36/
+1 0 mpl,if p(v435) then ,(v435^*,if (v32[1]='ahead' or v32[1]='review') and val(v65[1]*0.4)>2007 then v65[1]*0.4 fi,|v|v31[1],|s|v131[1],|n|v32[1],|s|v132[1],v41[1]/) fi
+1 0 mpl,if p(v935) then ,(v935,if (v32[1]='ahead' or v32[1]='review') and val(v65[1]*0.4)>2007 then v65[1]*0.4 fi,|v|v31[1],|s|v131[1],|n|v32[1],|s|v132[1],v41[1]/) fi
+2 0 mpl,|FSTUPD=|v991/
+5 0 mpl,if a(v200) or s(v200)='0' then 'MKPDONE=0'/ fi"
+6 0 mpl,v35/
+6 0 mpl,v930/
