+++ src/scielo/bin/xml/modules/xmlcvrter.py
@@ -2,26 +2,64 @@
 
 import os
 import shutil
+import tempfile
+from datetime import datetime
 
-from configuration import Configuration
 from isis_models import IssueRecord
-
-import files_manager
+import ftp_service
+import email_service
+import file_extractor
+import serial_files
 import reports
 import pkg_checker
 import xml_versions
+import isis
+import xmlcvter_cfg
 
-
-curr_path = os.path.dirname(__file__).replace('\\', '/')
 
 html_report = reports.ReportHTML()
 converter_report_lines = []
+CURRENT_PATH = os.path.dirname(__file__).replace('\\', '/')
 
 
 def register_log(message):
     if not '<' in message:
         message = html_report.format_message(message)
     converter_report_lines.append(message)
+
+
+def extract_package(pkg_file, pkg_work_path):
+    """
+    Extract files to pkg_work_path from compressed files that are in compressed_path
+    """
+    r = False
+
+    if os.path.isfile(pkg_file):
+        if file_extractor.is_compressed_file(pkg_file):
+            if not os.path.exists(pkg_work_path):
+                os.makedirs(pkg_work_path)
+            # delete content of destination path
+            if os.path.exists(pkg_work_path):
+                for item in os.listdir(pkg_work_path):
+                    os.unlink(pkg_work_path + '/' + item)
+            # create tempdir
+            temp_dir = tempfile.mkdtemp().replace('\\', '/')
+
+            # extract in tempdir
+            if file_extractor.extract_file(pkg_file, temp_dir):
+                # eliminate folders
+                for item in os.listdir(temp_dir):
+                    _file = temp_dir + '/' + item
+                    if os.path.isfile(_file):
+                        shutil.move(_file, pkg_work_path)
+                    elif os.path.isdir(_file):
+                        for f in os.listdir(_file):
+                            if os.path.isfile(_file + '/' + f):
+                                shutil.move(_file + '/' + f, pkg_work_path)
+                        shutil.rmtree(_file)
+                shutil.rmtree(temp_dir)
+                r = True
+    return r
 
 
 def get_valid_article(articles):
@@ -59,7 +97,7 @@ def convert_package(serial_path, xml_path, report_path, web_path, db_issue, db_a
     register_log(html_report.format_list('', 'ol', [os.path.basename(f) for f in xml_filenames]))
 
     for xml_filename in xml_filenames:
-        doc_files_info = files_manager.DocumentFiles(xml_filename, report_path, None)
+        doc_files_info = serial_files.DocumentFiles(xml_filename, report_path, None)
         doc_files_info.new_xml_filename = xml_filename
         doc_files_info_list.append(doc_files_info)
 
@@ -95,9 +133,9 @@ def convert_package(serial_path, xml_path, report_path, web_path, db_issue, db_a
             #register_log('Issue: ' + issue_label + '.')
             issue_isis = IssueRecord(issue_record)
             issue = issue_isis.issue
-            journal_files = files_manager.JournalFiles(serial_path, issue.acron)
-            ahead_manager = files_manager.AheadManager(db_ahead, journal_files, db_issue, issue.issn_id)
-            issue_files = files_manager.IssueFiles(journal_files, issue_label, xml_path, web_path)
+            journal_files = serial_files.JournalFiles(serial_path, issue.acron)
+            ahead_manager = serial_files.AheadManager(db_ahead, journal_files, db_issue, issue.issn_id)
+            issue_files = serial_files.IssueFiles(journal_files, issue_label, xml_path, web_path)
             issue_files.move_reports(report_path)
             issue_files.save_source_files(xml_path)
             report_path = issue_files.base_reports_path
@@ -112,6 +150,7 @@ def convert_package(serial_path, xml_path, report_path, web_path, db_issue, db_a
     print('\n\nXML Converter report:\n ' + converter_report_filename)
     pkg_checker.display_report(converter_report_filename)
     print('\n\n-- end --')
+    return report_path
 
 
 def stats(content=None):
@@ -250,95 +289,186 @@ def convert_article(db_article, issue_record, issue_files, xml_name, article, ah
     if article is not None:
         if ahead is not None:
             article._ahead_pid = ahead.ahead_pid
-        article_files = files_manager.ArticleFiles(issue_files, article.order, xml_name)
+        article_files = serial_files.ArticleFiles(issue_files, article.order, xml_name)
         r = db_article.create_id_file(issue_record, article, article_files)
     return r
 
 
-def has_xml_files(path):
-    files = [f for f in os.listdir(path) if os.path.isfile(path + '/' + f) and f.endswith('.xml')]
-    return len(files) > 0
+def delete_folder(folder):
+    if os.path.isdir(folder):
+        for item in os.listdir(folder):
+            if os.path.isfile(folder + '/' + item):
+                os.unlink(folder + '/' + item)
+            else:
+                delete_folder(folder + '/' + item)
+                shutil.rmtree(folder + '/' + item)
 
 
-def find_xml_paths(path):
+def execute_download_packages(ftp, ftp_dir, download_path, email, email_to, email_subject, email_header):
     r = []
-    if has_xml_files(path):
-        r.append(path)
-    else:
-        for d in os.listdir(path):
-            if os.path.isdir(path + '/' + d):
-                if has_xml_files(path + '/' + d):
-                    r.append(path + '/' + d)
+    log = ''
+    if ftp is not None:
+        if ftp_dir is not None and download_path is not None:
+            r = ftp.download_files(ftp_dir, download_path)
+            log = ftp.registered_actions()
+            if email is not None:
+                email_header = ''
+                if os.path.isfile(email_header):
+                    email_header = open(email_header, 'r').read()
+                email.send(email_to, email_subject, email_header + log)
+    return (r, log)
+
+
+def queue_packages(download_path, temp_path, queue_path, email, email_to, email_subject, email_header):
+    invalid_pkg_files = []
+    proc_id = datetime.now().isoformat()[5:10].replace('-', '')
+    temp_path = temp_path + '/' + proc_id
+    queue_path = queue_path + '/' + proc_id
+
+    if not os.path.isdir(temp_path):
+        os.makedirs(temp_path)
+
+    for pkg_file in os.listdir(download_path):
+        shutil.move(download_path + '/' + pkg_file, temp_path)
+
+    for pkg_file in os.listdir(temp_path):
+        pkg_name = os.path.basename(pkg_file)
+        pkg_path = queue_path + '/' + pkg_name
+        os.makedirs(pkg_path)
+
+        if not extract_package(temp_path + '/' + pkg_file, pkg_path):
+            invalid_pkg_files.append(pkg_file)
+            delete_folder(pkg_path)
+        os.unlink(temp_path + '/' + pkg_file)
+    delete_folder(temp_path)
+
+    if len(invalid_pkg_files) > 0 and email is not None:
+        email_header = ''
+        if os.path.isfile(email_header):
+            email_header = open(email_header, 'r').read()
+        email.send(email_to, email_subject, email_header + '\n'.join(invalid_pkg_files))
+    return queue_path
+
+
+def find_xml_source_paths(path):
+    r = []
+    if os.path.isdir(path):
+        files = [f for f in os.listdir(path) if os.path.isfile(path + '/' + f) and f.endswith('.xml')]
+        folders = [f for f in os.listdir(path) if os.path.isdir(path + '/' + f)]
+
+        if len(folders) > 0:
+            for item in folders:
+                r += find_xml_source_paths(path + '/' + item)
+        elif len(files) > 0:
+            r.append(path)
     return r
 
 
-def convert(path, config, version):
-    #FIXME
-    xml_paths = find_xml_paths(path)
-    if len(xml_paths) == 0:
-        print('XML files not found in ' + path + '\n')
-    elif len(xml_paths) > 0:
-        web_path = config.serial_path.replace('serial', '4web') if config.web_path is None else config.web_path
-        serial_path = config.serial_path
-
-        if open(curr_path + '/issue.fst', 'r').read() != open(config.issue_db + '.fst', 'r').read():
-            shutil.copyfile(curr_path + '/issue.fst', config.issue_db + '.fst')
-            config.isis_dao.update_indexes(config.issue_db, config.issue_db + '.fst')
-        issue_dao = files_manager.IssueDAO(config.isis_dao, config.issue_db)
-        article_dao = files_manager.ArticleDAO(config.isis_dao)
-
-        for xml_path in xml_paths:
-            print('*'*80)
-            print(xml_path + '\n')
-            report_path = xml_path + '_base_reports'
-            convert_package(serial_path, xml_path, report_path, web_path, issue_dao, config.isis_dao, article_dao, version)
-            print('*'*80)
-
-
-def read_configuration():
-    filename = curr_path + '/../../scielo_paths.ini'
-
+def xml_converter_read_configuration(filename):
+    r = None
     if os.path.isfile(filename):
-        r = Configuration(filename)
-        if r is not None:
-            if not r.valid():
-                r = None
-    else:
-        r = None
-    if r is None:
-        print(filename)
+        r = xmlcvter_cfg.XMLConverterConfiguration(filename)
+        if not r.valid():
+            r = None
     return r
 
 
-def read_inputs(args):
-    path = None
-    error_messages = ''
-    if len(args) == 2:
-        script, path = args
-        path = path.replace('\\', '/')
-        if not os.path.isdir(path):
-            path = None
+def xml_converter_read_inputs(args):
+    # python xml_converter.py <xml_src>
+    # python xml_converter.py <collection_acron>
+    messages = []
+    xml_paths = None
+    script = None
+    param = None
 
-    if path is None:
-        messages = []
+    if len(args) == 2:
+        script, param = args
+        if os.path.isdir(param):
+            param = param.replace('\\', '/')
+            if os.path.isdir(param):
+                xml_paths = find_xml_source_paths(param)
+            configuration_filename = CURRENT_PATH + '/../../cfg/scielo_paths.ini'
+
+        else:
+            configuration_filename = CURRENT_PATH + '/../config/' + param + '.xmlproc.ini'
+
+        if not os.path.isfile(configuration_filename):
+            messages.append('\n===== ATTENTION =====\n')
+            messages.append('ERROR: unable to read XML Converter configuration file: ' + configuration_filename)
+            param = None
+
+    if param is None:
         messages.append('\n===== ATTENTION =====\n')
         messages.append('ERROR: Incorrect parameters')
-        messages.append('\nUsage:')
-        messages.append('python xml_converter.py <xml_src>')
+        messages.append('\nUsages:')
+        messages.append('python xml_converter.py <xml_folder>')
         messages.append('where:')
-        messages.append('  <xml_src> = XML filenames folder')
-        error_messages = '\n'.join(messages)
+        messages.append('  <xml_folder> = path of folder which contains')
+        messages.append('or')
+        messages.append('python xml_converter <collection_acron>')
+        messages.append('where:')
+        messages.append('  <collection_acron> = collection acron')
 
-    return (path, error_messages)
+    else:
+        if xml_paths is not None:
+            if len(xml_paths) == 0:
+                messages.append('ERROR: Missing <xml folder>')
+    error_messages = '\n'.join(messages)
+
+    return (xml_paths, configuration_filename, error_messages)
+
+
+def update_issue_copy(issue_db, issue_db_copy):
+    d = os.path.dirname(issue_db_copy)
+    if not os.path.isdir(d):
+        os.makedirs(d)
+
+    if open(CURRENT_PATH + '/issue.fst', 'r').read() != open(issue_db_copy + '.fst', 'r').read():
+        shutil.copyfile(CURRENT_PATH + '/issue.fst', issue_db_copy + '.fst')
+    shutil.copyfile(issue_db + '.mst', issue_db_copy + '.mst')
+    shutil.copyfile(issue_db + '.xrf', issue_db_copy + '.xrf')
 
 
 def call_converter(args, version='1.0'):
-    config = read_configuration()
-    if config is None:
-        print('ERROR: Unable to configure XML Converter.')
+    messages = []
+    xml_paths, configuration_filename, error_messages = xml_converter_read_inputs(args)
+    if len(error_messages) > 0:
+        messages.append(error_messages)
     else:
-        path, error_messages = read_inputs(args)
-        if path is None:
-            print(error_messages)
-        else:
-            convert(path, config, version)
+        config = xml_converter_read_configuration(configuration_filename)
+
+        serial_path = config.serial_path
+        web_path = config.serial_path
+
+        isis_dao = isis.IsisDAO(isis.UCISIS(isis.CISIS(config.cisis1030), isis.CISIS(config.cisis1660)))
+
+        update_issue_copy(config.issue_db, config.issue_db_copy)
+        isis_dao.update_indexes(config.issue_db_copy, config.issue_db_copy + '.fst')
+        issue_dao = serial_files.IssueDAO(isis_dao, config.issue_db_copy)
+
+        article_dao = serial_files.ArticleDAO(isis_dao)
+
+        email_header = ''
+        if os.path.isfile(config.data.get('EMAIL_TEXT')):
+            email_header = open(config.data.get('EMAIL_TEXT'), 'r').read()
+
+        email = None
+        if xml_paths is None:
+            ftp = ftp_service.FTPService(config.data.get('FTP_SERVER'), config.data.get('FTP_USER'), config.data.get('FTP_PASSWORD'))
+            if config.data.get('EMAIL_SERVICE') == 'on':
+                email = email_service.EmailService(config.data.get('SENDER_NAME'), config.data.get('SENDER_EMAIL'))
+
+            execute_download_packages(ftp, config.data.get('FTP_DIR'), config.data.get('DOWNLOAD_PATH'), email, config.data.get('EMAIL_TO'), config.data.get('EMAIL_SUBJECT_DOWNLOAD'), config.data.get('EMAIL_TEXT_DOWNLOAD'))
+            queue_path = queue_packages(config.data.get('DOWNLOAD_PATH'), config.data.get('TEMP_PATH'), config.data.get('QUEUE_PATH'), email, config.data.get('EMAIL_TO'), config.data.get('EMAIL_SUBJECT_NOT_PROCESSED'), config.data.get('EMAIL_TEXT_NOT_PROCESSED'))
+            xml_paths = [item for item in os.listdir(queue_path)]
+
+        for xml_path in xml_paths:
+            messages.append('*'*80)
+            messages.append(xml_path + '\n')
+            report_path = xml_path + '_base_reports'
+            report_path = convert_package(serial_path, xml_path, report_path, web_path, issue_dao, isis_dao, article_dao, version)
+            messages.append('*'*80)
+
+            if email is not None:
+                email_text = open(report_path + '/xml_converter_report.html', 'r').read()
+                email.send(config.data.get('EMAIL_TO'), config.data.get('EMAIL_SUBJECT'), email_header + email_text, attaches=os.listdir(report_path))
